relocate `Rotation` declaration & put it in `io` namespace
