+++ src/grid/grid3D.h
@@ -50,73 +50,11 @@
   #include "../analysis/analysis.h"
 #endif
 
-struct Rotation {
-  // TODO: refactor so we can remove the default constructor
-  Rotation() = default;
-
-  /*! primary constructor */
-  Rotation(const Parameters &P);
-
-  /*! \var nx
-   *   \brief Number of pixels in x-dir of rotated, projected image*/
-  int nx;
-
-  /*! \var nz
-   *   \brief Number of pixels in z-dir of rotated, projected image*/
-  int nz;
-
-  /*! \var nx_min
-   *   \brief Left most point in the projected image for this subvolume*/
-  int nx_min;
-
-  /*! \var nx_max
-   *   \brief Right most point in the projected image for this subvolume*/
-  int nx_max;
-
-  /*! \var nz_min
-   *   \brief Bottom most point in the projected image for this subvolume*/
-  int nz_min;
-
-  /*! \var nz_max
-   *   \brief Top most point in the projected image for this subvolume*/
-  int nz_max;
-
-  /*! \var delta
-   *   \brief Rotation angle about z axis in simulation frame*/
-  Real delta;
-
-  /*! \var theta
-   *   \brief Rotation angle about x axis in simulation frame*/
-  Real theta;
-
-  /*! \var phi
-   *   \brief Rotation angle about y axis in simulation frame*/
-  Real phi;
-
-  /*! \var Lx
-   *   \brief Physical x-dir size of projected image*/
-  Real Lx;
-
-  /*! \var Lz
-   *   \brief Physical z-dir size of projected image*/
-  Real Lz;
-
-  /*! \var i_delta
-   *   \brief number of output projection for delta rotation*/
-  int i_delta;
-
-  /*! \var n_delta
-   *   \brief total number of output projection for delta rotation*/
-  Real n_delta;
-
-  /*! \var ddelta_dt
-   *   \brief rate of delta rotation*/
-  Real ddelta_dt;
-
-  /*! \var flag_delta
-   *  \brief output mode for box rotation*/
-  int flag_delta;
-};
+// forward-declare the Rotation struct
+namespace io
+{
+struct Rotation;
+}
 
 struct Header {
   /*! \var n_cells
@@ -510,12 +448,12 @@ class Grid3D
   /*! \fn void Write_Header_Rotated_HDF5(hid_t file_id)
    *  \brief Write the relevant header info to the HDF5 file for rotated
    * projection. */
-  void Write_Header_Rotated_HDF5(hid_t file_id, Rotation &R);
+  void Write_Header_Rotated_HDF5(hid_t file_id, io::Rotation &R);
 
   /*! \fn void Write_Rotated_Projection_HDF5(hid_t file_id)
    *  \brief Write rotated projected data to a file, at the current simulation
    * time. */
-  void Write_Rotated_Projection_HDF5(hid_t file_id, const Rotation &R);
+  void Write_Rotated_Projection_HDF5(hid_t file_id, const io::Rotation &R);
 
   /*! \fn void Write_Slices_HDF5(hid_t file_id)
    *  \brief Write xy, xz, and yz slices of all data to a file. */
