+++ src/io/RotatedProjWriter.cpp
@@ -11,7 +11,7 @@
 #include "../grid/grid3D.h"
 #include "../io/io.h"
 
-Rotation::Rotation(const Parameters &P)
+io::Rotation::Rotation(const Parameters &P)
 {
 #ifdef ROTATED_PROJECTION
   // x-dir pixels in projection
